+++ src/api/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to
 
 ### Changed
 
+- Update the list of active operational units
+
 #### Dependencies
 
 - Upgrade `alembic` to `1.16.5`