+++ src/api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.30.0] - 2025-10-31
+
 ### Changed
 
 - Update the list of active operational units
@@ -612,7 +614,8 @@ update` command
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.29.0...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.30.0...main
+[0.30.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.29.0...v0.30.0
 [0.29.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.28.0...v0.29.0
 [0.28.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.27.0...v0.28.0
 [0.27.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.26.0...v0.27.0